Title: 배열의 특정 위치 찾기, Time: 98ms, Memory: 8MB, Status: Passed - CodeTree

+++ 241025/배열의 특정 위치 찾기/find-specific-location-fo-array.java
@@ -11,7 +11,7 @@ public class Main {
 
         int sum =0;
 
-        int avg = 0;
+        double avg = 0;
 
         for(int i=1; i<=10; i++){
             int value = Integer.parseInt(st.nextToken());
